client/requirements.txt
Pillow~=11.2.0
aiohttp~=3.11.0
black~=25.1.0
displayhatmini~=0.0.2
python-dotenv~=0.9.0
